refactor(advanced): type 위치 변경
