--- src/advanced/domain/product.ts
@@ -1,9 +1,0 @@
-export type Product = {
-  id: string;
-  name: string;
-  discountPrice: number;
-  price: number;
-  quantity: number;
-  onSale: boolean;
-  suggestSale: boolean;
-};
